Changed first arguement (Fill)
Changed the  secound argument of the fill method from 0 to 125
Expectation: that the colour of the elipse would change in some way.
Outcome: The colour did change but only in brightness. Going from black to whi
Outcome: The colour did change but only in brightness/Shade going from dark
Outcome: The colour did change but only in brightness/Shade going from dark
Outcome: The colour did change but only in brightness/shade
on the spectrum from black to white i left it at grey